rework: changed login functions from provider to where its actually needed
